server/requirements.txt
# AI Fitness Monitor Server - Dependencies
# =========================================

# Web Framework
Flask>=2.2.5,<3.0.0
gunicorn>=21.0.0,<22.0.0

# Computer Vision
opencv-python>=4.8.0,<5.0.0
mediapipe>=0.10.9,<0.11.0

# Numerical Processing
numpy>=1.24.0,<2.0.0

# Text-to-Speech
pyttsx3>=2.90,<3.0.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0

# Development
python-dotenv>=1.0.0,<2.0.0
